raksha-backend/requirements.txt
fastapi 
uvicorn
sqlalchemy 
databases 
pydantic
mysqlclient
aiomysql
pyjwt
passlib
